dev-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index dev-requirements.in
#
appdirs==1.4.3            # via black
asgiref==3.2.3
astroid==2.3.3            # via pylint
attrs==19.3.0             # via black
black==19.10b0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.0                # via black
django-cors-headers==3.2.1
django-filter==2.2.0
django==3.0.3
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3  # via astroid
markdown==3.2.1
mccabe==0.6.1             # via pylint
pathspec==0.7.0           # via black
phonenumbers==8.11.4
pycparser==2.19
pylint==2.4.4
pynacl==1.3.0
pytz==2019.3
regex==2020.2.20          # via black
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
telnyx==1.0.0b2
toml==0.10.0              # via black
typed-ast==1.4.1          # via astroid, black
urllib3==1.25.8
wrapt==1.11.2             # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
